tests/wpt/tests/html/canvas/element/manual/draw-element-image/privacy/visited-link-decoration-color-ignored.tentative.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElementImage does not reveal visited link colors</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src="/html/canvas/resources/wait-for-canvas-paint.js"></script>
  <style>
    #link {
      color: green;
      text-decoration: underline 20px solid green;
      width: 100px;
      height: 100px;
      display: block;
    }
    #link:visited {
      color: red;
      text-decoration: underline 20px solid red;
    }
  </style>
</head>
<body>
  <canvas id="canvas" width="100" height="100" layoutsubtree>
    <a id="link" href="">link</a>
  </canvas>

  <script>
    window.onload = () => {
      promise_test(async function(t) {
        await waitForCanvasPaint(canvas);

        var context = canvas.getContext("2d");

        // Draw the element image
        context.drawElementImage(link, 0, 0);

        // The canvas should NOT reflect the visited state (red), it should paint the unvisited state (green).
        let pixel = context.getImageData(20, 40, 1, 1).data;
        assert_false(pixel[0] > 0, "The canvas should not reflect the visited state.");
        assert_true(pixel[1] > 0, "The canvas should not reflect the visited state.");
      });
    };
  </script>
</body>
</html>
